requirements.txt
flask
mysql-connector-python
qrcode
pillow
openpyxl
